Add files via upload
The biothings file is in fact the biothings.api file (I changed the name so that it can be imported).
The config.py file is the one I have copied and modified from the biothings documentation p.7.
config_hub.py and conf.py comes from the biothings.api folder. 
dump.py is the script containing BioMart you gave me a bit earlier.
dumper.py is one I am currently working on for the plant dumper.